Edit questions and answers inline of the festival
There are two ways to manage questions and answers. Either using the
dedicated `Questions` tab from the admin panel or inline of the festival
details. Nothing changed for the `Questions` tab but when editing
questions inline of the festival the festival is preconfigured and any
questions that are created are already created for that particular
festival.