Fix the creation of inline questions
Not sure how it slipped but creating new questions inline of festivals
didn't work. The issue was that the `festivalId` wasn't set and
therefore failed schema validation server side when making the PUT
request. This is an easy fix by setting the `festivalId` as an hidden
input field. This makes sure that the field is included in the HTTP
request.